Make driver more configurable, clean, and explicit
The status errors were moved into an enum to be consistent and make it
easier to decode external of the a02yyuw class.

The serialdata tuple was created to make it easier to keep track of what
each byte means.

The only public methods needed for quickstart are now read, status, and
distance. Read reads from the sensor to update the status and distance
states.

Overall, these changes aim to bring more configurability and clarity to
the module.